Refactored the rest of the K means implementation to use 3 dimensional points

+++ Point3D.cpp
@@ -1,7 +1,7 @@
 #include "Point3D.h"
 #include <math.h>
 
-Point3D::Point3D(int x, int y, int z) : 
+Point3D::Point3D(double x, double y, double z) : 
     Point(x, y), _z(z)
 {
 }
@@ -10,7 +10,7 @@ Point3D::~Point3D()
 {
 }
 
-double Point3D::distance( int pos_x, int pos_y, int pos_z ) 
+double Point3D::distance( double pos_x, double pos_y, double pos_z ) 
 {
     return(sqrt(pow((getXPos() - pos_x), 2) + pow((getYPos() - pos_y), 2)) + pow((getZPos() - pos_z), 2));
 }
